quality of life changes

+++ Photo_Prestiges/ExterneService/index.js
@@ -199,21 +199,26 @@ app.get('/uploadtarget', opaqueTokenCheck, async function(req, res, next) {
 
 
 app.listen(port, async() => {
-    await connectToRabbitMQ();
+    console.log('Externe is up on port ' + port);
     
-    await consumeFromQueue('imageDataResponseQueue', '', 'image_data_response', async (data, dbname) => {
-        await db.collection('uploadtargets').insertOne(data);
-    });
-
-    await consumeFromDirectExchange("targetDeleteExchange", "users", "delete_target_from_externe_service", async (data, dbname) => {
-        await uploadTargetModel.deleteOne({tid: data.tid});
-        console.log(`Removed target with tid ${data.tid} from uploadtarget collection`);
-    });
-
-    await consumeFromDirectExchange("targetDeleteExchange", "users", "delete_target_from_user_externe_service", async (data, dbname) => {
-        await uploadTargetModel.deleteOne({tid: data.tid});
-        console.log(`Removed target with tid ${data.tid} from uploadtarget collection`);
-    });
-
-    console.log('Server is up on port ' + port);
+    if(await connectToRabbitMQ() == false) {
+        console.log("RabbitMQ is not connected");
+        res.json({message: "RabbitMQ is not connected"});
+    } else {
+        await connectToRabbitMQ();
+    
+        await consumeFromQueue('imageDataResponseQueue', '', 'image_data_response', async (data, dbname) => {
+            await db.collection('uploadtargets').insertOne(data);
+        });
+    
+        await consumeFromDirectExchange("targetDeleteExchange", "users", "delete_target_from_externe_service", async (data, dbname) => {
+            await uploadTargetModel.deleteOne({tid: data.tid});
+            console.log(`Removed target with tid ${data.tid} from uploadtarget collection`);
+        });
+    
+        await consumeFromDirectExchange("targetDeleteExchange", "users", "delete_target_from_user_externe_service", async (data, dbname) => {
+            await uploadTargetModel.deleteOne({tid: data.tid});
+            console.log(`Removed target with tid ${data.tid} from uploadtarget collection`);
+        });
+    }
 });
